Allow existing areas that are created to pass validation
Instead of throwing an error when the entered created area already exists, allow this to pass validation, since this is a better user experience. This is also p
ossible because the server is looking up the area and only creating a new area i
f it doesn't already exist. The logic of adding the area to the metadata list wa
s also updated by checking that the area is not already in the metadata areas, which is what is used to populate the dropdown.