Rename docker deployment script
This mirrors the name of our .travis_version_bump script from the
mark-one project. This commit also tweaks the logic for determining the
docker tag, replacing the conditional variable assignment with a switch
statement that also exits on any other branches.

I also added a call to "docker logout" to the end, just in case.