Set timeout for two minutes
With the timeout completely disabled, there was one test run on Travis
that stopped testing at the integration tests, but still treated it as
passing. I re-ran the build a couple times and it seemed to work, but
I'm going to set the timeout for two minutes instead, because I think a
false negative result from the test timing out is more useful than a
false positive from the test never running at all.