Log into docker before build
Similar to the previous commit, we end up needing to pull down the
latest version of the node image when we build our app image and this
can also be subject to rate limiting. So we're moving the docker login
and logout steps out of the docker_deploy script entirely, and running
them with the before_script and after_deploy commands.

Note that each Travis job runs on a clean environment, so we can't
persist the login between jobs, so we have to duplicate the login and
logout in both test and build.