Update BadRequestExceptionPipe to remove condensing of error messages into a string and update Course Admin Modal to use server side validation
On submission of the Course Admin Modal, a POST or PUT request will be done depending on whether the user is creating or editing a course. Then it will take the error messages, if any, from the server, and I've set it up to convert the camel case version of the property (e.g. termPattern, isSEAS, courseTitle, etc.) to a more user friendly and readable version (e.g. Term Pattern, Is SEAS, Course Title). The BadRequestExceptionPipe was updated so that we can use the property and restraint properties to identify the field with the error along with the accompanying error message. The property provided by the server helps us identify which error message slot to accordingly put the error message in.