Update test to use a mockdb due to transition to server side validation
Because we are now doing server side validation, we need to use a mockdb instead of stubbing out the results of the database. The POST and PUT stubs for courses were replaced with calling the mockdb so that we can check that the onSuccess and onClose stubs are called or not called appropriately. I also removed the client side validation tests, since we are doing all of the validation on the server now.