Update tests to check response body differently now that BadRequestExceptionPipe has been updated
Originally, the BadRequestExceptionPipe was flattening all of the error messages into one string. The BadRequestExceptionPipe was updated so that when there is a validation error, it returns an array of objects. Each of the objects contains a property that reflects the field that has an issue, e.g. termPattern. The tests were updated to reflect the new structure of the response by looping through this array and checking to see if one of the properties matches the field that we are searching for.