Fix Faculty Controller tests to account for change in BadRequestExceptionPipe
Originally, the BadRequestExceptionPipe was returning errors by concatenating the error messages into one long string. Now that the response is structured differently, the tests checking for the error field needed to check in a different way. To do this, we are looping through the response's error property names and checking that the field with the issue is included in this array.