Fix callsFake call to send the course data and add additional tests to check modal error output
The callsFake call was originally being called wrong, which is why we were not able to call the server. In order to transfer the course data, we must use the '.send' method. Also, supertest does not throw the error for you, so I needed to throw the error myself. I created a class, AxiosSupertestError, which takes a supertest response with an error and reshapes it into the same structure as an AxiosError. I also added tests to check the behavior of all the fields in the Course Admin modal to check that the messages do and do not show up appropriately when field data is missing or when the type of data provided is wrong (in the cases where term pattern and Is SEAS expect their respective enum values).