Clean up database after each test runs
The original Population Service tests were not properly closing their
connections to the MockDB, which is:

- Not practicing good isolation, as all of the depopulation tests
depend on the population tests to run first
- Breaking the MockDB for any tests that run after the Population
Service

We've been dodging that second bullet for a while, but the integration
tests for the Course Admin Modal in feature/229-course-admin-modals are
running into it and need it to be fixed.

In addition to closing the direct connection from TypeORM after each
depopulation test runs, I'm changing the `before`s and `after`s that
create/stop the nest module to `beforeEach`s and `afterEach`s so that
each inidividual test runs against a clean database, which addresses the
concern from  bullet 1 above.