Add functional filters for catalog prefix, catalog number, and course title
The filters for the first three fields were added. For now, the catalog prefix dropdown is being populated by the metadata areas array, but this will need to be replaced with a list of the catalog prefixes, which are currently not yet in the metadata. The function, filteredMultiYearPlans, was added to make the filters functional.