Skip cleanup on build deploy
Includes some other small tweaks:

- replace old-style `skip_cleanup: true` with `cleanup: false`, as this
will be changing in a future version
- Calling the docker deploy script explicitly with `bash`

The second point is attempting to address an issue where our builds on
develop are still failing on initial run (though they seem to be working
in debug mode on Travis). It still can't find the remote repo, even with
the updated GH token, so it may be an environment-related issue.