Rewrite Documentation workflow in raw git
Replaces the gh-pages command used previously, which did not work
correctly, with git commands that accomplish the same thing.

Looking at the code[1] for the gh-pages library, it seems that gh-pages
is cloning a fresh copy of the repo into a cache directory, then
attempting to push the docs to gh-pages from there. In the process, it's
losing the credential configuration that sets up the github token to
provide push access, so that's  why it isn't working.

[1]: https://github.com/tschaub/gh-pages/blob/main/lib/index.js