Checkout branch HEADs, instead of refs
By default, the checkout action will checkout only the single commit
that kicked off the workflow process. We actually want to get the
reference to the HEAD of the particular branch, so that we can get the
latest commits that would have been pushed in by our "Increment Version"
workflow.