Compare main/develop versions
Because our Docker workflow is triggered by the completion of the
versioning workflow instead of a push/pr event, it will always run in
the context of our default branch, main. This means that a change to the
version on our develop branch won't be reflected in the package.json on
main yet. So, we'll need to grab the versions from main and develop
separately, then compare them with the semver tool and set only the
latest one as our VERSION_TAG variable. Then we can use that variable to
checkout just the specific version tag to be built.