Update multi-year GET controller to account for case of insufficient semester rows in db
In the case in which we are requesting more semesters than exist in the database, we would like to return the number of semesters requested by generating empty semester objects. This commit adds a function, createMissingSemester, to do so. It calculates what the term, academic year, and calendar year should be based off of what semesters were returned by the server and how many semesters were requested.