Transfer logic of supplementing missing semesters to the client instead of the server
Originally, the controller was supplementing the missing semesters. Instead, we decided to have the client handle it to better separate the logic between client and server. The controller was reverted to its original state of returning the data as its received from the server. The client will now populate the year headers and filters based on the constant NUM_SEMESTERS that was added to the 'common/constants' file... meaning that regardless of what is returned from the server, the correct number of columns in the multiyear plan table will be rendered.