Call the getSemestersFromYear function once instead of calling it multiple times
Instead of calling the getSemestersFromYear function multiple times in different places, I called it once and stored it in a semesters variable to avoid recalculation. Also, I used the getSemestersFromYear function in the courseInstance definition instead of constructing the area (with Array.from...). With getSemestersFromYear, I can use the key that is calculated in calculateSemesters.