Check for existence of room data before rendering
Originally the code assumed that if there was meeting data for an
instance in a given semester, then there would also be room data. This
is incorrect and doesn't match the data we're pulling in from the old
course planner. Now we're checking to see if we have room data before
attempting to render it.