Return a single array of instances
Previously we had intended to allow for fetching multiple academic Years
worth of data at a time, but A) that doesn't actually seem necessary for
the MVP and B) this structure was causing some issues when there is no
data for a given year. Instead we can just fetch one year at a time, and
send multiple requests from the client instead.