Create IsMutuallyExclusiveWith validator
Adds a custom validator to validate cases where a request should contain
exactly one from among a set of fields. This is specifically to handle
cases where a meeting must be associated with either a CourseInstance or
a NonClassEvent, but not both, though it may be useful in other
circumstances.