Create a CoursesPageContext to pass course instance information to
The course instance table is dynamically generated using functions. In order to more cleanly pass the course information without altering the existing functions by adding more parameters, this context was added to be used to set the current course instance and appropriately assign the ref on modal open and close.