Move utils.ts into utils/ directory
It could be a little confusing to have a 'util.ts' file and a 'utils/'
directory with an 'index.ts' inside (making both '../util' and
'../utils' valid import sources). We don't have a consistent rule for
when there should be a file or folder, but because the
`IsMutuallyExclusiveWith` validator was a more complicated chunk of code
I thought it needed its own file and, thus, should be in a subdirectory.

So I've moved the `trimString` function into a transformers.ts file and
exported all named exports from the index.ts. If we end up creating more
and more of these validators/transformers/etc, we may want to split out
a separate npm package for them so we can bring them into other
projects as needed.