Add test to make sure that the original edit meeting button is focused
Codecov is indicating that a path has not been covered in that we are not checking the case when the course id is the same but the term is not the same. In this case, we want to check that when we originally clicked the fall edit button for a certain course that the fall button gets focused when the modal closes and not the spring button of this particular course.