Create RoomBookingInfoView
This view brings in information specific to when and why rooms are
booked, including the year, term, day, startTime, and endTime. It only
includes the id of the room itself, but can be joined with the
RoomListingView on "roomId" to bring in that data as well.

This was originally created to be used in validation for ticket 304 (the
PUT method for meetings) but is being split off into a separate PR as
it is also applicable to ticket 307 (the GET method for rooms)