Treat empty results as "available"
If there's a room in the database with no associated bookings, the query
might return no reults at all. In that case, we should treat the room as
available.

This commit also provides a type for the return value of the query, and
adds some additional inline comments to better explain what's happening.