Remove parent ids from DTOs
Since a single request to the api will potentially contain multiple
meetings, it doesn't make sense to have the id of the associated
non-class meeting or course instance duplicated on each of the meetings.
Instead, the parent ID will be part of the request path.

This obviates the IsMutuallyExclusiveWith validator, but I've left in as
it may be useful in other scenarios.