Cascade updates to meetings
In order to be able to remove/update existing meetings, we'll actually
need to save the entire course instance or non-class event in our
meeting service, which requires setting the option to cascade updates.

We're also setting the orphanedRowsAction in the meetings, so that when
a meeting is removed from the parent entity it will be deleted, not left
as a null entry in the database.