Update typeorm to 0.2.31
The orphanedRowAction option is included in a new version of typeorm,
but in order to use that we needed to make changes to our log service to
fix the function signature for errors.

Along the way, I realized that our LogService wasn't correctly being
configured as a "Global" service, as the `@Global()` decorator needs to
live on the Module, not the service itself.