Consolidate logic in the meeting service
This splits the logic of validating and creating/updating the meeting
entities and saving them to the parent into separate functions, one of
which calls the other. This gives us slightly clearer flow of data, and
consolidates our database calls under the service rather than having
some in the controller.