Recreate ScheduleEntryView without JOINs
This migration implements the changes from the previous two commits by
dropping the existing ScheduleEntryView and recreating it without the
JOIN statements.

I did need to manually tweak the timestamp so that this migration would
be run before 1618340375982-NonClassEventViews.ts, as the change in that
migration would not run without this change being made first. This means
that anyone who has already run 1618340375982-NonClassEventViews.ts will
need to either:

1) run migration:revert to undo it.
2) wipe out the postgres container and re-run all migrations.

I recommend option 2