Add RoomService
The room service joins the RoomListingView with the RoomBookingInfoView, and it gets room data along with meetings that are occurring within the requested time periods. The "where" clause for start time and end time covers the cases in which the meeting spans before and after the requested start and end times, the meeting is within the requested start and end times, the meeting starts before the requested time and ends within the requested time, and lastly, the meeting starts within the requested times and ends after the requested times.