Update service to utilize leftJoinAndMapMany's parameters
The where clauses were not filtering the data correctly, as the desired result for this service is to return all rooms - the previous version was returning one item in the array. The query was also returning the columns from the RoomBookingView as b_columnname (e.g. b_term, etc.). By transferring the where statements into leftJoinAndMapMany's and removing the meetings property via the mapping statement at the bottom of the service method, we are able to get the desired response( RoomResponse[]).