Specify explicit Session types
This resolves some of the type issues that have frequently popped up
around the session and the fields listed in it. The SessionData
interface is specifically meant to allow for adding custom fields with
declaration merging:

https://github.com/DefinitelyTyped/DefinitelyTyped/blob/master/types/express-session/index.d.ts#L202-L215

The inconsistency of when these errors appear seems to have something to
do with the resolution of the express, express-session, and
connect-redis packages and their associated @types packages. I've added
some more up-to-date version strings for those, which should at least
ensure the same packages are pulled regularly.