Create unique meetings for nonClassEvents
The testing process for the meeting controller revealed an issue with
the way that nonClassEvents were being populated. It was pulling 4
meetings from the database and assigning them to new nonClassEvents, but
all of the existing meetings in the database are from the course
population process so they're already associated with a course. This is
technically legal from the database perspective, since nonClassEventId
and CourseInstanceId are separate fields, but it means the controller
tests were failing when we try to edit meetings since the rooms would
show up as being booked by the existing course Instance.

To resolve this, we need to reformat the data to embed the meeting
details with the other event details, then create the new meetings
within the map function that's creating the new nonClassEvents.

This also changes the assignment of semesters to ensure that version of
the non-class meetings exist in all semesters, which is how the course
instances are populated and thus closer to what our true implementation
will be like.