Test the meeting controller with nonClassEvents
This uses the test data now being populated in the mock database to
ensure that the PUT method in the meeting controller successfully
creates new meetings in the system.

This also removes some stubbed tests related to authentication
permissions, since doing so would mainly be testing the testingStrategy
rather testing the actual code.