Add room controller integration tests
These check the behavior unauthenticated/authenticated/admin/non-admin users. Authenticated users should have access, and a valid term, day, start time, and end time must be provided. If the user enters a year that we do not have data for, all of the rooms are still returned but with no meeting titles.