Replace date parsing with regex
The parsing of timezones in the tests was causing different values to be
computed depending on the timezone of the system running the tests. This
broadly seems to be the limitation of trying to do time-based
computation with Date functions in Javascript. In its place, this commit
adds a new tzStringToAMPM function that parses the zoned string with a
regex and reformats it into AM/PM time.

Note: This commit might be obviated by #327