Improve exception handling
This will keep legitimate errors from being swallowed, and replaces the
NotFoundError, which is an HTTP error and doesn't belong at the database
layer, with an EntityNotFoundError, which does. This is a slight hack of
the EntityNotFoundError constructor[1], which normally expects to
receive an EntityTarget as its first argument. Since we're throwing this
error explicitly to indicate that neither a courseInstance nor a
nonClassEvent matches, we're passing this object instead.

[1]: https://github.com/typeorm/typeorm/blob/f85f436f51fb000cd9959b44e8d7a79bf0cd10ab/src/error/EntityNotFoundError.ts