Fix table alias in expression
This may have been a rebase/merge issue at some point, but the
ScheduleEntryView expression is trying to use "c" as an alias for the
campus table, but it's already been used for "course". This change
already exists in the 1618340375981-RemoveViewDependencies migration
(hence why I suspect it's a merge issues) and that's why it keeps trying
to create the ScheduleEntryView over again.