Create Migration to change column types
This is a ghastly migration, but I couldn't figure out a better way to
handle the views' dependency on the startTime and endTime. I did try to
manually edit the migration to use an ALTER COLUMN statement instead of
the drop/add method, but that still ran into the dependency problem so
we would have had to drop and recreate the views doing it that way as
well.