Add tests to check calculateStartEndTimes helper function
This commit moves the calculateStartEndTimes from the constants directory to the common utils directory and adds tests to make sure the helper function is working correctly. When provided a timeslot string (e.g. 11:00 AM - 12:00 PM), the helper function should return the start time and end time, which are used to populate the start and end time input boxes.