Add validateTimes function
The validateTimes function checks the existing meeting being edited to make sure each of the fields are filled out and that the start date is not later than or equal to the end date. If the time input field is only partially filled out, its value is an empty string. The validateTimes function returns true if there are no errors - this keeps the user from being able to switch to a different meeting while there's an error in the current row.