Update node image to 14
Node v10 is now officially end of life, so we're updating to the current
LTS version, 14.

This also drops the `-stretch` from the tag so that we use the default
version of the container instead -- currently Debian 10 (Buster) -- as
Debian 9 (Stretch) is on "unofficial LTS" status while Buster is still
active.