Update nest, typeorm, and pg
Possibly this is only an issue on my system, but while running node v14
with our previous versions of nest, typeorm, and pg, I was unable to
connect to the postgres container so I could not run the tests or the
ETL.

My suspicion (after several hours of debugging) is that something in the
EventEmitter class changed and the way that pg was attempting to create
a socket connection to postgres was hanging and never completing the
connection. Whether that's true or not, updating the packages seems to
fix the problem for me. fingers crossed it stays fixed.

However, Nest v7 includes some breaking changes to the way that Errors
are returned from the server[1]. This is not a bad thing, as it
reconciles inconsistencies between different exceptions, but some of our
tests are based on the inconsistent exceptions, so more commits will
follow to fix those tests.

[1]: https://docs.nestjs.com/migration-guide#http-exceptions-body