Restore legacy behavior of ValidationPipe
The ValidationPipe behavior was changed in Nest v7, tranforming the
array of ValidationErrors into an array of strings containing the
validation messages. This is less useful for our use case, as the
validation messages no longer have the context of which field created
the errors. Following the advice in Nest's Migration Guide[1], we can
restore the existing behavior for now.

Possibly in the future, we could write a more customized
exceptionFactory that would provide a happier medium between the two
styles.

[1]: https://docs.nestjs.com/migration-guide#validation-errors-schema