Fix CourseAdmin integration tests
I don't think this is directly related to the update to nest, but was
flagged as a failing test because the structure of the
InternalServerException changed in the update. In both test cases, the
axios put/pull methods were being stubbed to return
InternalServerExceptions directly. This is incorrect, as 1) the client
request method shouldn't throw a server exception and 2) the test is
meant to determine have the client code reacts to an error on the
server.

The updated code uses `callsFake` to send the data payload to the
supertestApi then stub out the underlying repository/service methods and
forces them to throw errors.