Add @types/vinyl-fs
This should be a temporary workaround to a bug in @nestjs/core, where
the declaration from vinyl-fs in their build environment is getting
wrapped into the distributed package. I've submitted an issue on the
repo (nestjs/nest#7247), and if it gets fixed we can revert this commit,
since @types/vinyl-fs is not needed here.