Fix convert12To24HourTime helper function
When the time input in the Meeting Modal is changed, it sets the state of the time on each key press. Because of the way that the convert12To24HourTime helper function was set up, every time 12 was entered in for the hour, it changed the period to "AM." This made it impossible to change the time to 12:00PM. Padding was also added to AM times so that the result is in hh:mm format.