Check for the absence of error messages
Check the innerText of the error message when clicking "Show Rooms." Since the row does not collapse when clicking this button, there is technically an empty span where the error message would go if there was one. Instead of checking for the number of elements with the role of alert, we are instead checking the innerText to make sure there is no error message.