Add function to "Add New Time" button
On Click of the Add New Time button, a new meeting should be added to the current list of meetings if there are no outstanding errors from other rows. A new meeting can only be added if all fields of the currently existing meetings are filled out and if the start time of the current meeting is earlier than the end time. To create a unique id for the new meeting, a state containing an index called "newMeetingIndex" was created. After each new meeting is created, this index is increased by 1. Presumably after the server request, a permanent UUID will be created for each of the newly created meetings to replace this temporary index-generated ID.