Remove trailing space in the case that there is no meeting room defined
In the case where meeting room is null, we do not want a trailing space after the meeting end time. This commit adds a meeting to the testCourseInstance in which the room is null to make sure that the text for each meeting row prints correctly whether the room is null or defined.