Update CSS of Meeting Modal
This commit separates the location onto a separate line when displaying the meeting times for a course instance. It also adds a minimum height to each row so that an error message does not cause a row to expand in height. A margin was also added around the "Show" and "Close" buttons. Padding was removed from each list item element so that the ends of the "Show Rooms" and "Add New Times" button could align.