Fix formatting of meeting times list items
Set a minimum width for each of the form fields so that when the window's width is decreased that the form fields are still visible. This also updates the grid layout to more accurately reflect the mockup with the form fields not stretching to the full width of the row - this helps the day dropdown and times slot dropdown sit a little closer together. A max width is set on the day dropdown so that it does not stretch to the entire width of the grid area.