Parse number in alpha and integer fields
Uses the event listener feature of TypeORM to parse the number field
passed into any update/insert event and split the values into the
numberInteger and numberAlphabetical fields. This is done with a regex
that matches the whole number field to [NUMBER][LETTERS] (e.g. 109a)
with both groups being optional.