Add index on prefix/integer/alphabetical
This adds a compound index to the table on the combination of prefix,
numberInteger, and numberAlphabetical, and adds the relevant CREATE
INDEX command to the migration.

Just to node: I did run some EXPLAINs in psql with and without the
index, and afaict it's not actually improving the query time on the
courses table itself. I think this is somewhat due to the fact that
there's a relatively small number of courses in the system, but it's
worth revisiting in the future if performance in the app is slow.