Move conditional rendering out of table
The original component design did all of the conditional rendering (e.g.
the loading spinner) inside in RoomSelectionTable, but that meant moving
additional state back and forth. Now the RoomSelection component handles
the data fetching state on its own, while the RoomSelectionTable handles
the availability selection on its own.

It may be necessary to move the room fetching action further up the
stack as we expand the left <==> right connections.