Change initial value of roomlist
Using `null` as the initial value was throwing an error in the browser
because our table component tries to call `.map()` on null, which is not
a thing. Because we're using an empty array for that now, we need
another way to decide when to show the "Select a time" prompt, which in
this case is now whether or not we have any room request data in props.