Move the state of the current meetings into the parent component, MeetingModal
The state of the current meetings was moved into the meeting modal, because we need to have a way for the left and right sides of the meeting modal to be able to interact with each other such that each side "knows" what the current meetings are. Instead of having the state of meetings within MeetingTimesList (the child), the state is now in the parent (MeetingModal). I created an onChange prop so that the state of the currentMeetings could be updated from MeetingTimesList. onChange is called when the user clicks an edit button, the close button, the show rooms button, and the add times button. The function "getCurrentMeetings" is used to retrieve and is used to update the current list of meetings from MeetingModal.