Move MeetingModal inside the Meeting cell
The original implementation used a Context wrapper around the Courses
page to pass around handlers and values to control the display of the
MeetingModal. This is functional, but it creates a lot of redirection
and spreads the implementation across many different component. This can
lead to strange bugs since it's creating essentially another source of
state beyond what's in our components themselves. (See React's docs for
[when not to use context][1])

This recomposes the component tree by moving the MeetingModal component
to be part of the formatMeetings method that renders the contents of the
Meetings Cell. Since we're implementing the modal with a Portal, the DOM
elememnt created by the Modal will always live at the bottom of the body
and will only ever be rendered if the meeting button is clicked, so we
shouldn't run the risk of rendering duplicates in a long list, though it
may be worth doing some profiling to see if this is causing delays for
any other reasons.

One other piece of data that was being passed through Context is the
calendar year of the term being edited. This data lives at the top
CoursesPages level, so we needed to add a second parameter to the
getValue function to pass that into the method for rendering the
meetingList. We may need to pass additional props down in this way in
the future, so it may be worth expanding that second paramater to be a
Partial<Record<string, unkown>> object so we can essentially pass named
arguments into the rendering function.

[1]: https://reactjs.org/docs/context.html#before-you-use-context