Move editing state into the MeetingModal
Having the values of the field be set by state in the MeetingTimesList
was going to cause problems when we need to update the value of the room
field from the RoomList on the righthand side of the modal. This changes
all of the value fields in the MeetingTimesList to read from a
currentEditMeeting state field in the MeetingModal. It does not change
the handlers used to update thsoe fields.