Move logic for closing/opening into MeetingModal
This moves the validateTimes function into the parent component, and
creates a single "closeCurrentEditMeeting" method that:

1. runs the validation
 a. If there are error, bail out and pass them into the List
 b. If its clean, merge the edits into allMeetings
2. If a newMeeting is passed as a parmeter, switch to that meeting as
the currentEditMeeting
3. If that newMeeting doesn't exist, merge it into allMeetings

This should cover all the use cases for the different button handlers
inside the meeting modal, including the Save button (since we'll want to
validate/merge any active changes there too).

It does involve a fair amount of conditional logic -- If it were any
more complex we might want to look at using a reducer instead and
defining differet actions for each situation.