Modify room search to match time format
The server expects a time string resolved to the second, not the minute
like our current version. This tweaks the regex in PGTime to make the
seconds field optional so that we can use it to convert our second-less
time, and adds a toRequestString that leaves out the milliseconds.

We will almost certainly tweak this further in the work on #358, but
this should suit our needs for now.