Update "Save" button tests
Previously, the save button would leave the current meeting open, so we
could test for the absence of an alert with no text as a passing
condition. Now that the save button closes the meeting, we need to check
for the absence of an alert to confirm that there are no errors.

Thanks to some fantastic work by Natalyn on this test suite, these were
the only tests that needed to be updated after refactoring, and that's
because they're the only ones where the observable behavior is changing.