Remove invalid condition
The only legitimate case for a meeting not being part of the allMeetings
array is when we're creating new one, which is initially passed in as
the value to `newMeeting` so it will be caught under the `if
(newMeeting)` block. If the currentEditMeeting isn't in the allMeetings
array already it's almost certainly an error and we should not merge it
in.