Change type of addRoomHandler
The original implementation was done while building the
RoomSelectionTable in isolation, so the type of the addRoom function was
assumed to use the same RoomResponse data we got from the server. This
is actually incorrect as the CourseInstanceResponseMeeting type is
what's used to display the data, and that type's room field doesn't
include the capacity or meetingTitles values.