Pass down addRoomHandler
The addRoomHandler just uses the same updateCurrentEditMeeting method
used for editing the meeting data on the left-hand side, but explicitly
updates the room field with whatever is passed in. One advantage of this
is that it already handles the case of blanking out the
RoomSelectionTable when the button is clicked.