Add excludeParent search option
This provides a way to pass the id of the parent of the meeting to the
room search so that we can exclude it from appearing as a match in the
meetingTitles response field.

This is done on the server side so we can match on the UUID and confirm
that it's exactly the same instance. The alternative would have been to
filter the meetingTitles on the client side, but that can't guarantee
accuracy since the title field we have may not be absolutely unique.
This also has the advantage of working for both courseInstances and
nonClassEvents.