Blank the "Show Rooms" data when modal appears
Adds one more handler to the useEffect hook that I forgot in c12abcb0.

While testing some other features, I realized that if you perform a
search for a room, then close the modal before selecting a room, then
re-open the modal on the same course-instance, the RoomSelectionTable
will automatically repopulate with the data from that last search.