Move "clear" operations to close phase
Rather then clear old data when the modal opens, we need to clear some
data when the modal closes. This was showing up as a bug where sometimes
the request for room data would still kick off when the modal opened,
because the state data was launching the request before the call to
clear it.