Add handler to remove meetings when delete button is clicked
The handler removeMeeting is passed down to MeetingTimesList so that the appropriate meeting can be deleted when the delete (trash can) icon is clicked. On click, if the meeting deleted was also at the time being edited, the state of the currentEditMeeting is set back to null. The state of the list of all meetings lives on the MeetingModal, and this state is updated on delete icon click as well.