Refactor getRooms for optional excludeParent
In order to have an optional search parameter in the query, we need to
use a subquery and only add the `exludeParent` check if there is
actually a value. (I initially tried to have excludeParent default to
either '' or null, but neither of those were compatible with the <>
operator when comparing against a UUID).