Ensure that a unique meeting id is generated each time a new meeting is added
The previous implementation depended on the length of allMeetings to generate an id number that was unique. However, the number of meetings can fluctuate back to the same id number while editing meetings in the meeting modal. Managing the id number via state ensures that the number generated to make the new meeting id is unique.