Replace .filter with within().queryAllByRole()
The usage of getAllbyRole().filter seemed to have a cumulative timeout
effect where the tests would pass when run in isolation with .only, but
when run as part of the full test suite would then timeout and fail.

The solution was to use within() to narrow down the search space of the
test function, making the test run faster and preventing the timeout
issue.

More info on this here: https://github.com/testing-library/dom-testing-library/issues/552