Flip boolean logic for table body rows
The changes made in 0d2ec50bca8852864cb3fd95c327b736cef88274
accidentally duplicated the logic from the header tests above and missed
the subtlety of the `!` to invert the logic.

This meant that the two tests were effectively testing the same thing,
and the test that was _claiming_ to test the number of rows in the table
body, was _actually_ the number of rows in the table head. Clearly this
is incorrect, but the test passed anyway because the number of rows of
table data(2) was the same as the number of rows of headings (also 2).

This commit adds another row of table (and test) data to ensure that
this cannot happen again.